+++ packages/liveblocks-react/src/__tests__/useMarkInboxNotificationAsRead.test.tsx
@@ -27,7 +27,7 @@ describe("useMarkInboxNotificationAsRead", () => {
     const inboxNotifications = [
       dummyThreadInboxNotificationData({
         roomId,
-        threadId: threads[0].id,
+        threadId: threads[0]!.id,
         readAt: null,
       }),
     ];
@@ -76,10 +76,10 @@ describe("useMarkInboxNotificationAsRead", () => {
 
     // Mark the first thread in our threads list as read
     act(() => {
-      result.current.markInboxNotificationAsRead(inboxNotifications[0].id);
+      result.current.markInboxNotificationAsRead(inboxNotifications[0]!.id);
     });
 
-    expect(result.current.inboxNotifications![0].readAt).not.toBe(null);
+    expect(result.current.inboxNotifications?.[0]?.readAt).not.toBe(null);
 
     unmount();
   });
@@ -90,7 +90,7 @@ describe("useMarkInboxNotificationAsRead", () => {
     const inboxNotifications = [
       dummyThreadInboxNotificationData({
         roomId,
-        threadId: threads[0].id,
+        threadId: threads[0]!.id,
         readAt: null,
       }),
     ];
@@ -139,15 +139,15 @@ describe("useMarkInboxNotificationAsRead", () => {
 
     // Mark the first thread in our threads list as read
     act(() => {
-      result.current.markInboxNotificationAsRead(inboxNotifications[0].id);
+      result.current.markInboxNotificationAsRead(inboxNotifications[0]!.id);
     });
 
     // We mark the notification as read optimitiscally
-    expect(result.current.inboxNotifications![0].readAt).not.toBe(null);
+    expect(result.current.inboxNotifications?.[0]?.readAt).not.toBe(null);
 
     await waitFor(() => {
       // The readAt field should have been updated in the inbox notification cache
-      expect(result.current.inboxNotifications![0].readAt).toEqual(null);
+      expect(result.current.inboxNotifications?.[0]?.readAt).toEqual(null);
     });
 
     unmount();